+++ simplemdm_sdk/_connection.py
@@ -22,6 +22,12 @@ class Connection:
         self._session      = requests.Session()
         self._session.auth = HTTPBasicAuth(api_key, "")
 
+        self._session.headers.update({
+            "Accept":       "application/json, application/x-plist",
+            "Content-Type": "application/json",
+            "User-Agent":   "SimpleMDM-SDK-Python / 0.1.0"
+        })
+
 
     def __epoch_to_seconds_remaining(self, epoch: int) -> int:
         """