+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Resolve `variable_support` error in `SimpleMDM.scripts.list_all()`.
 - Resolve `managed_app_configs.py` URL error.
 - Resolve `script_jobs.py`'s `create` method failing to build array of identifiers.
+- Set `User-Agent` header to `SimpleMDM-SDK-Python / {version}`.
+- Set `Accept` header to `application/json, application/x-plist`
+- Set `Content-Type` header to `application/json`
 
 ### 0.0.1
 - Initial alpha release.